refactor: move the Bash helpers into src/lib modules
Groundwork for the PreToolUse coverage work stacked on top of this. That
change adds a shell tokenizer, around forty command classifications and a
tool-input scanner; landing all of it in pre-tool-use-hook.ts would take the
file past 1400 lines and give it several unrelated reasons to change.

The two questions are separated here, while each answer is still small
enough to read in one sitting:

- src/lib/shell.ts — shell syntax, with nothing about what a command means.
- src/lib/bash-commands.ts — what a command does with its operands.

Nothing is rewritten. Of the 45 non-blank lines removed from the hook, 44
appear verbatim in the new modules; the one that does not is the section
banner the move left empty. develop's 343 tests pass untouched, which is
what shows the behaviour did not move with the code.
